Prototype implementation of epoll() support
Bulk patch

git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-epoll-kw-20101115@8864 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ resip/stack/MsgHeaderScanner.cxx
@@ -942,7 +942,7 @@ MsgHeaderScanner::scanChunk(char * chunk,
       localTextPropBitMask |= charInfo->textPropBitMask;
      determineTransitionFromCharCategory:
       TransitionInfo *transitionInfo =
-         &(localStateMachine[localState][(size_t)charCategory]);
+         &(localStateMachine[(unsigned)localState][(size_t)charCategory]);
       TransitionAction transitionAction = transitionInfo->action;
 #if defined(RESIP_MSG_HEADER_SCANNER_DEBUG)  
       printStateTransition(localState, *charPtr, transitionAction);
